requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==0.8.0
aiohttp==3.7.4
aiormq==3.3.1
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.8.3
APScheduler==3.7.0
argh==0.26.2
asn1crypto==1.3.0
astropy==4.0.1.post1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bidict==0.21.4
bkcharts==0.2
bleach==3.1.5
blis==0.7.6
boto==2.49.0
boto3==1.21.10
botocore==1.24.10
Bottleneck==1.3.2
brotlipy==0.7.0
CacheControl==0.12.10
cached-property==1.5.2
cachetools==4.2.4
catalogue==2.0.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
chatette==1.6.3
clang==5.0
click==7.1.2
cloudpickle==1.6.0
clyent==1.2.2
colorama==0.4.4
colorclass==2.2.2
coloredlogs==10.0
colorhash==1.0.4
comtypes==1.1.7
constantly==15.1.0
contextlib2==0.6.0.post1
contextvars==2.4
cryptography==36.0.1
cycler==0.10.0
cymem==2.0.6
cytoolz==0.10.1
decorator==4.4.2
defusedxml==0.6.0
dm-tree==0.1.6
dnspython==1.16.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
enum-compat==0.0.3
et-xmlfile==1.0.1
fastcache==1.1.0
fbmessenger==6.0.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
flatbuffers==1.12
fsspec==0.7.4
future==0.17.1
gast==0.4.0
gevent==1.5.0
glob2==0.7
gmpy2==2.0.8
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.44.0
h11==0.9.0
h5py==3.1.0
HeapDict==1.0.1
httpcore==0.11.1
httptools==0.4.0
httpx==0.15.4
huggingface-hub==0.0.12
humanfriendly==10.0
hyperlink==21.0.0
idna-ssl==1.1.0
imagesize==1.2.0
incremental==21.3.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
json5==0.9.5
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.1.5
kafka-python==2.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
klein==17.10.0
langcodes==3.3.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.33.0+1.g022ab0f
locket==0.2.0
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.3.4
mattermostwrapper==2.2
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.0.4
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.4.0
mpmath==1.1.0
msgpack==1.0.0
multidict==5.0.0
multipledispatch==0.6.0
murmurhash==1.0.6
nbconvert==5.6.1
nbformat==5.0.7
nose==1.3.7
notebook==6.0.3
npm==0.1.1
numba==0.50.1
numexpr==2.7.1
numpy==1.19.5
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
optional-django==0.1.0
packaging==20.9
pamqp==2.3.0
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
preshed==3.0.6
prometheus-client==0.8.0
prompt-toolkit==2.0.10
protobuf==3.19.4
psutil==5.7.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycrypto==2.6.1
pycurl==7.43.0.5
pydantic==1.8.2
pydot==1.4.2
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==2.3.0
pykwalify==1.8.0
pymongo==3.10.1
pyodbc===4.0.0-unsupported
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pyTelegramBotAPI==3.8.3
pytest==5.4.3
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
QtPy==1.9.0
questionary==1.9.0
rasa==2.8.25
rasa-nlu==0.15.1
rasa-sdk==2.8.4
redis==3.5.3
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rocketchat-API==1.16.0
rope==0.17.0
rsa==4.8
Rtree==0.9.4
ruamel.yaml.clib==0.2.6
ruamel_yaml==0.15.87
s3transfer==0.5.2
sacremoses==0.0.47
sanic==20.9.1
Sanic-Cors==0.10.0.post3
sanic-jwt==1.6.0
Sanic-Plugins-Framework==0.9.5
scikit-image==0.16.2
scikit-learn==0.24.2
seaborn==0.10.1
Send2Trash==1.5.0
sentry-sdk==1.2.0
simplegeneric==0.8.1
simplejson==3.17.6
singledispatch==3.4.0.3
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
sniffio==1.2.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
spacy==3.0.8
spacy-alignments==0.8.4
spacy-loggers==1.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
srsly==2.4.2
statsmodels==0.11.1
tables==3.6.1
tabulate==0.8.9
tarsafe==0.0.3
tblib==1.6.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-probability==0.13.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.10
testpath==0.4.4
thinc==8.0.13
tokenizers==0.10.3
toolz==0.10.0
torch==1.10.2
tornado==6.0.4
traitlets==4.3.3
transformers==4.9.2
twilio==6.50.1
Twisted==22.1.0
twisted-iocpsupport==1.0.2
typed-ast==1.4.1
typeguard==2.13.3
typing==3.7.4.3
typing-extensions==3.10.0.2
tzlocal==2.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.25.11
webencodings==0.5.1
webexteamssdk==1.6
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwings==0.19.5
xlwt==1.3.0
yarl==1.7.2
zict==2.0.0
zipp==3.1.0
zope.event==4.4
zope.interface==4.7.1
